10 excepciones

+++ API/src/main/java/com/svalero/API/service/ClientService.java
@@ -2,6 +2,7 @@ package com.svalero.API.service;
 
 import com.svalero.API.domain.Client;
 import com.svalero.API.domain.dto.ClientDto;
+import com.svalero.API.exceptions.ClientNotFoundException;
 import com.svalero.API.repository.ClientRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,7 +21,10 @@ public class ClientService {
         return this.clientRepository.findAll();
     }
 
-    public Client getById(long id){
+    public Client getById(long id) throws ClientNotFoundException {
+        Client client = this.clientRepository.findById(id);
+        if(client == null)
+            throw new ClientNotFoundException("Client with ID " + id + " not found");
         return this.clientRepository.findById(id);
     }
 
@@ -29,14 +33,18 @@ public class ClientService {
         return this.clientRepository.save(client);
     }
 
-    public int remove(long id){
-        //this.clientRepository.findById(id);
+    public int remove(long id) throws ClientNotFoundException {
+        Client client = this.clientRepository.findById(id);
+        if(client == null)
+            throw new ClientNotFoundException("Client with ID " + id + " not found");
         this.clientRepository.deleteById(id);
         return 1;
     }
 
-    public Client modify(long id, ClientDto updatedClient){
+    public Client modify(long id, ClientDto updatedClient) throws ClientNotFoundException {
         Client client = this.clientRepository.findById(id);
+        if(client == null)
+            throw new ClientNotFoundException("Client with ID " + id + " not found");
         modelMapper.map(updatedClient, client);
         return this.clientRepository.save(client);
     }
